Updated setuprdp.sh only prompt for Xclients customization for XFCE if startxfce4 is in the PATH

+++ installers/setuprdp.sh
@@ -17,18 +17,20 @@ if egrep -i -q 'centos|fedora|oracle|red *hat' /etc/*release* 2>/dev/null ; then
 
     if egrep -i -q '(oracle|red *hat).* [67]' /etc/*release* 2>/dev/null ; then
 
-      # Xclients customization in /etc/X11/xinit for XFCE:
-      # TODO verify if applies to centos & fedora
-      echo ${BASH_VERSION:+-e} "Force Xclients to startxfce4? [y/N]\c"
-      read ans
-      if [ "$ans" = y ] ; then
-        sudo mv -v /etc/X11/xinit/Xclients /etc/X11/xinit/Xclients.bak.$(date '+%Y-%m-%d-%OH%OM%OS')
-        cat <<'EOF' | sudo tee /etc/X11/xinit/Xclients
+      if which startxfce4 >/dev/null 2>&1 ; then
+        # Xclients customization in /etc/X11/xinit for XFCE:
+        # TODO verify if applies to centos & fedora
+        echo ${BASH_VERSION:+-e} "Force Xclients to startxfce4? [y/N]\c"
+        read ans
+        if [ "$ans" = y ] ; then
+          sudo mv -v /etc/X11/xinit/Xclients /etc/X11/xinit/Xclients.bak.$(date '+%Y-%m-%d-%OH%OM%OS')
+          cat <<'EOF' | sudo tee /etc/X11/xinit/Xclients
 #!/usr/bin/env bash
 exec $(which startxfce4)
 EOF
+        fi
+        sudo chmod 755 /etc/X11/xinit/Xclients
       fi
-      sudo chmod 755 /etc/X11/xinit/Xclients
 
     fi
   fi
